refactor the code.
